multiple-viewports without using instances
In this attempt at rendering multiple-viewports I am 
not using a library node and instances, instead the
nodes rendered in each renderer node are basically 
a duplicate of the hello-teapot scenegraph.

Rendered I only see the second (upper-left) viewport.

If I comment out just final teapot node in the second renderer 
instance the first teapot appears in the upper-left.

{
    // type : "teapot"
}