reset canvas size for webgl when specified with css
When the size of a canvas element is set with css styles
instead of specifying the size in the element itself the
canvas displays using the css styles but reports a default
height and width (300 x 150).

When creating a webgl context the canvas attributes height
and width are used so if the canvas size is specified with
a css style the resulting webgl viewport will be incorrect.